Added social network events to gui.
Created generic event panel class to display histories both of votes and of follows.  Created RecentSocials DTO object to return a list of recent follows for an editor.

Eliminated transient fields from editions and users.  Now a separate DTO is explicitly required to be constructed if you want to populate an object graph and send it to the client.