Implemented Log in, Log out functions, created a Vote module to replace VoteServlet, implement show voters list for each story, and add a support this story checkbox to each story
For login: A log in link is now in the top left corner of the RNA app.  This fetches a login url from the server and redirects to the google service at that url, which then redirects back to the rna app.  every user action now triggers a check with the server for the login state, which is updated accordingly on the client side.  (This check is necessary because a login may be invoked by the new vote module and we need a way to update the display on the RNA module).

Voting is now a separate gwt module, which requires a login.

Story records now contain a checkbox for supporting the story, which invokes voting code on the server.

A class called Voters has been added to the StoryRecord class which requests a list of voters voting for the story, and displays them by hooking into the Open event fired by the Disclosure Panel (show votes).